refactored load
