+++ libs/langchain/langchain/base_language.py
@@ -1,4 +1,4 @@
-"""Deprecated module for BaseLanguageModel class, kept for backwards compatibility."""
+"""DEPRECATED: Kept for backwards compatibility."""
 from __future__ import annotations
 
 from langchain.schema.language_model import BaseLanguageModel
