+++ libs/langchain/langchain/chains/llm.py
@@ -14,7 +14,6 @@ from langchain.callbacks.manager import (
     Callbacks,
 )
 from langchain.chains.base import Chain
-from langchain.load.dump import dumpd
 from langchain.prompts.prompt import PromptTemplate
 from langchain.schema import (
     BaseLLMOutputParser,
@@ -24,6 +23,7 @@ from langchain.schema import (
     StrOutputParser,
 )
 from langchain.schema.language_model import BaseLanguageModel
+from langchain.utils.dump import dumpd
 from langchain.utils.input import get_colored_text
