+++ libs/langchain/langchain/example_generator.py
@@ -1,4 +1,4 @@
-"""Keep here for backwards compatibility."""
+"""DEPRECATED: Kept for backwards compatibility."""
 from langchain.chains.example_generator import generate_example
 
 __all__ = ["generate_example"]
