+++ libs/langchain/langchain/load.py
@@ -1,0 +1,23 @@
+"""DEPRECATED: Kept for backwards compatibility."""
+from langchain.utils.dump import default, dumpd, dumps
+from langchain.utils.load import Reviver, loads
+from langchain.utils.serializable import (
+    Serializable,
+    SerializedConstructor,
+    SerializedNotImplemented,
+    SerializedSecret,
+    to_json_not_implemented,
+)
+
+__all__ = [
+    "Reviver",
+    "Serializable",
+    "SerializedConstructor",
+    "SerializedNotImplemented",
+    "SerializedSecret",
+    "default",
+    "dumpd",
+    "dumps",
+    "loads",
+    "to_json_not_implemented",
+]
