+++ libs/langchain/langchain/python.py
@@ -1,4 +1,4 @@
-"""For backwards compatibility."""
+"""DEPRECATED: Kept for backwards compatibility."""
 from langchain.utilities.python import PythonREPL
 
 __all__ = ["PythonREPL"]
