+++ libs/langchain/langchain/schema/memory.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import Any, Dict, List
 
-from langchain.load.serializable import Serializable
 from langchain.schema.messages import AIMessage, BaseMessage, HumanMessage
+from langchain.utils.serializable import Serializable
 
 
 class BaseMemory(Serializable, ABC):
