+++ libs/langchain/langchain/schema/retriever.py
@@ -5,10 +5,10 @@ from abc import ABC, abstractmethod
 from inspect import signature
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
-from langchain.load.dump import dumpd
-from langchain.load.serializable import Serializable
 from langchain.schema.document import Document
 from langchain.schema.runnable import Runnable, RunnableConfig
+from langchain.utils.dump import dumpd
+from langchain.utils.serializable import Serializable
 
 if TYPE_CHECKING:
     from langchain.callbacks.manager import (
