+++ libs/langchain/langchain/serpapi.py
@@ -1,4 +1,4 @@
-"""For backwards compatibility."""
+"""DEPRECATED: Kept for backwards compatibility."""
 from langchain.utilities.serpapi import SerpAPIWrapper
 
 __all__ = ["SerpAPIWrapper"]
