+++ libs/langchain/langchain/sql_database.py
@@ -1,4 +1,4 @@
-"""Keep here for backwards compatibility."""
+"""DEPRECATED: Kept for backwards compatibility."""
 from langchain.utilities.sql_database import SQLDatabase
 
 __all__ = ["SQLDatabase"]
