+++ libs/langchain/langchain/utils/__init__.py
@@ -4,6 +4,7 @@
 These functions do not depend on any other LangChain module.
 """
 
+from langchain.utils.dump import default, dumpd, dumps
 from langchain.utils.env import get_from_dict_or_env, get_from_env
 from langchain.utils.formatting import StrictFormatter, formatter
 from langchain.utils.input import (
@@ -12,7 +13,15 @@ from langchain.utils.input import (
     get_colored_text,
     print_text,
 )
+from langchain.utils.load import Reviver, loads
 from langchain.utils.math import cosine_similarity, cosine_similarity_top_k
+from langchain.utils.serializable import (
+    Serializable,
+    SerializedConstructor,
+    SerializedNotImplemented,
+    SerializedSecret,
+    to_json_not_implemented,
+)
 from langchain.utils.strings import comma_list, stringify_dict, stringify_value
 from langchain.utils.utils import (
     check_package_version,
@@ -24,11 +33,19 @@ from langchain.utils.utils import (
 )
 
 __all__ = [
+    "Reviver",
+    "Serializable",
+    "SerializedConstructor",
+    "SerializedNotImplemented",
+    "SerializedSecret",
     "StrictFormatter",
     "check_package_version",
     "comma_list",
     "cosine_similarity",
     "cosine_similarity_top_k",
+    "default",
+    "dumpd",
+    "dumps",
     "formatter",
     "get_bolded_text",
     "get_color_mapping",
@@ -37,10 +54,12 @@ __all__ = [
     "get_from_env",
     "get_pydantic_field_names",
     "guard_import",
+    "loads",
     "mock_now",
     "print_text",
     "raise_for_status_with_text",
     "stringify_dict",
     "stringify_value",
+    "to_json_not_implemented",
     "xor_args",
 ]
