+++ libs/langchain/langchain/utils/dump.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any, Dict
 
-from langchain.load.serializable import Serializable, to_json_not_implemented
+from langchain.utils.serializable import Serializable, to_json_not_implemented
 
 
 def default(obj: Any) -> Any:
