+++ libs/langchain/tests/unit_tests/load/test_dump.py
@@ -8,10 +8,10 @@ from langchain.callbacks.tracers.langchain import LangChainTracer
 from langchain.chains.llm import LLMChain
 from langchain.chat_models.openai import ChatOpenAI
 from langchain.llms.openai import OpenAI
-from langchain.load.dump import dumps
-from langchain.load.serializable import Serializable
 from langchain.prompts.chat import ChatPromptTemplate, HumanMessagePromptTemplate
 from langchain.prompts.prompt import PromptTemplate
+from langchain.utils.dump import dumps
+from langchain.utils.serializable import Serializable
 
 
 class Person(Serializable):
